Add files via upload
This is the illustration of the output of the coding. 
At first we already got the output but then we want to change some features of the game so we change the coding and 
after that the output doesn't come out. So this the attachment of our imagination output should be :